services/cancel-order-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pika==1.3.2
python-dotenv==1.0.1
